added two new html files and a javascript file
added 3 files, two new html files i.e create-articles.html and navbar.html and a JavaScript file extend-navbar.js. In any new page all we need to do to extend the navbar is just to add a div or section that makes use of the "navbar" class. the JavaScript "extend-navbar" will automatically include the navbar